update clang version

+++ receiver/src/data_cache.cpp
@@ -17,7 +17,7 @@ DataCache::DataCache(uint64_t cache_size, float keepunlocked_ratio) : cache_size
     }
 
     srand(static_cast<unsigned int>(time(NULL)));
-    counter_ = rand() % 100 + 1;
+    counter_ = static_cast<uint32_t>(rand() % 100 + 1);
 }
 
 void* DataCache::AllocateSlot(uint64_t size) {
